celeryApp/requirements.txt
celery
requests
confluent_kafka 
psutil
python-dotenv
redis
httpx